Updated error handling and added regression test scripts
Moved error handling utilities from loggers to utils, introducing a KwexError exception for improved error management.

Added try/except/finally structure to main workflow for improved cleanup and error logging.

Added regression-test scripts for automated regression testing.

+++ src/kwex/main.py
@@ -9,25 +9,28 @@ from .velocity import *
 
 def kwex_script(args):
     """ Function that contains the kwex workflow. """
-    init_log()
-    compile_velocity()
+    try:
+        init_log()
+        compile_velocity()
 
-    product_list = get_files(args.input, args.input_root)
-    pointer_list = get_pointers(args.template)
-    write_nows_template(run_state.nows_template_str)
-    init_spice(pointer_list)
-    init_velocity(run_state.nows_template_filename)
-    
-    for product in product_list.values():
-        output_path = get_output(args.output, product)
-        val_list = get_values(product, pointer_list)
-        send_values(val_list, output_path)
+        pointer_list = get_pointers(args.template)
+        product_list = get_files(args.input, args.input_root)
+        
+        write_nows_template(run_state.nows_template_str)
+        init_spice(pointer_list)
+        init_velocity(run_state.nows_template_filename)
+        
+        for product in product_list.values():
+            output_path = get_output(args.output, product)
+            val_list = get_values(product, pointer_list)
+            send_values(val_list, output_path)
 
-        if get_config(ConfigKey.DATA_OUTPUT) == ConfigState.MOVE:
-            product.data_product.unlink()
+            if get_config(ConfigKey.DATA_OUTPUT) == ConfigState.MOVE and output_path.exists():
+                product.data_product.unlink()
 
-    terminate_velocity()
-    log_bad_output()
-    run_state.nows_template_filename.unlink()
-    if run_state.temp_kernel:
-        run_state.kernel_path.unlink()
+        log_bad_output()
+    except KwexError as e:
+        logger.get_logger().error(e.message)
+    finally:
+        terminate_velocity()
+        cleanup()
